feat(styles): add global.scss and main.scss files
The global.scss file is added to define global styles for the application. It sets padding and margin to 0, uses the border-box box-sizing model, and sets the font-family to "Montserrat" for all elements.

The main.scss file is added and imports the global.scss file and the responsive/index.scss file. This allows for the use of global styles and responsive styles throughout the application.